Update containerd to 1.5.9.
This updates github.com/containerd/containerd to 1.5.9 to resolve
the CVE noted here:
https://github.com/allenai/beaker/security/dependabot/1

To prepare these changes I ran:

```
go get -u github.com/containerd/containerd
go mod tidy
```

To test I build a local `beaker` binary and ran a few commands.
This was in no way thorough.

I looked for tests, but it doesn't look like we have any.

I'm not too worried about breaking anything. We don't actually
use the CRI runtime (or `containerd`) in the client. It's included
because we use Beaker's runtime package to attach to Docker
containers that've been launched by the executor. So it's really
unlikely that the containerd version shift will cause problems.